Add files via upload
Snow’s Adventure:
o An adventure game made using Gamemaker
o This is a character-based game showcasing Snow White as the main character.
o To win in this game, Snow White has to collect all the mushrooms and crystals she can find on her way to
the level exit sign at each level. The collectibles are worth different points.
o The monsters in this game are giant flies, hydras, dragons, and goblins. If Snow runs into the giant flies or
the hydras, she will die and the current level will restart. If she runs into a goblin, she earns 100 pts. Also, if
Snow hits the dragon, it transforms into a goblin.